ai-agent-system/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
python-dotenv==1.1.0
groq==0.23.1
tavily-python==0.5.1
google-generativeai==0.8.5
httpx==0.28.1
Pillow==11.2.1
PyPDF2==3.0.1
pandas==3.0.2
sambanova
